Adds event bus
