+++ lib/hud.dart
@@ -19,10 +19,10 @@ class Hud extends PositionComponent with HasGameReference<RTSGame> {
 
   @override
   void update(double dt) {
-    if (game.world.highlightedWorld == null) {
+    if (game.eventBus.highlightedWorld == null) {
       hudElement.text = 'Selected: None';
     } else {
-      hudElement.text = 'Selected: ${game.world.highlightedWorld}';
+      hudElement.text = 'Selected: ${game.eventBus.highlightedWorld}';
     }
     return;
   }
